dom/indexedDB/test/test_globalObjects_chrome.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<window title="Mozilla Bug 832883"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="runTest();">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <script type="application/javascript">
  <![CDATA[
  function* testSteps() {
    const name = window.location.pathname;

    // Test for IDBKeyRange and indexedDB availability in chrome windows.
    var keyRange = IDBKeyRange.only(42);
    ok(keyRange, "Got keyRange");

    var request = indexedDB.open(name, 1);
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    let event = yield undefined;

    let db = event.target.result;
    ok(db, "Got database");

    finishTest();
    yield undefined;
  }
  ]]>
  </script>

  <script type="text/javascript" src="chromeHelpers.js"></script>

  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=832883"
     target="_blank">Mozilla Bug 832883</a>
  </body>
</window>
